Add docker image for the app
The docker file requires an update for the service that runs on
container creation.
I changed the .net framework version for YaqeenDAL project to 7.0 to
allow a successful build for the solution.
Also, I fixed a small typo in services word and added a fix for one of
the props as it was preventing a successful build.